DMP-465 - update after review
