+++ src/main/java/uk/gov/hmcts/darts/cases/CasesRoute.java
@@ -3,6 +3,6 @@ package uk.gov.hmcts.darts.cases;
 import com.service.mojdarts.synapps.com.GetCases;
 import com.service.mojdarts.synapps.com.GetCasesResponse;
 
-public interface CasesService {
+public interface CasesRoute {
     GetCasesResponse getCases(GetCases getCasesRequest);
 }
