+++ src/main/java/uk/gov/hmcts/darts/cases/impl/CasesRouteImpl.java
@@ -4,23 +4,23 @@ import com.service.mojdarts.synapps.com.GetCases;
 import com.service.mojdarts.synapps.com.GetCasesResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.darts.cases.CasesService;
+import uk.gov.hmcts.darts.cases.CasesRoute;
 import uk.gov.hmcts.darts.cases.mapper.GetCasesMapper;
-import uk.gov.hmcts.darts.common.client.ModernisedDartsFeignClient;
+import uk.gov.hmcts.darts.common.client.DartsFeignClient;
 import uk.gov.hmcts.darts.model.cases.ScheduledCase;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class CasesServiceImpl implements CasesService {
+public class CasesRouteImpl implements CasesRoute {
 
-    private final ModernisedDartsFeignClient modernisedDartsFeignClient;
+    private final DartsFeignClient dartsFeignClient;
 
     @Override
     public GetCasesResponse getCases(GetCases getCasesRequest) {
 
-        List<ScheduledCase> modernisedDartsResponse = modernisedDartsFeignClient.getCases(
+        List<ScheduledCase> modernisedDartsResponse = dartsFeignClient.getCases(
             getCasesRequest.getCourthouse(),
             getCasesRequest.getCourtroom(),
             getCasesRequest.getDate()
