--- src/main/java/uk/gov/hmcts/darts/ws/CasesApi.java
@@ -1,28 +1,0 @@
-package uk.gov.hmcts.darts.ws;
-
-import com.service.mojdarts.synapps.com.GetCases;
-import com.service.mojdarts.synapps.com.GetCasesResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.ws.server.endpoint.annotation.Endpoint;
-import org.springframework.ws.server.endpoint.annotation.PayloadRoot;
-import org.springframework.ws.server.endpoint.annotation.RequestPayload;
-import org.springframework.ws.server.endpoint.annotation.ResponsePayload;
-import uk.gov.hmcts.darts.cases.CasesService;
-
-@Endpoint
-@RequiredArgsConstructor
-@Slf4j
-public class CasesApi {
-
-    private final CasesService casesService;
-
-    private final DartsResponseUtils utils;
-
-    @PayloadRoot(namespace = "http://com.synapps.mojdarts.service.com", localPart = "getCases")
-    @ResponsePayload
-    public GetCasesResponse getCases(@RequestPayload GetCases getCases) {
-        return casesService.getCases(getCases);
-    }
-
-}
